Remove obsolete kernel builder module and finish consistentcy testing
